added custom distance metric toolbar; some minor clean up
This patch adds a toolbar with a combo box for specifying units of measure. The default is meters, but other choices include inches, yards, feet, cm. Also, the current measurement can be used to define a new metric, e.g., Smoots.

A few misc. clean ups in activity.py as well introduced while I was trying to sort out issues caused by the spam bug and the launching problem.

Signed-off-by: Rafael Ortiz <rafael@activitycentral.com>